requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
